Fix ubuntu18.04 gcc 7.5 (#17)
* Fix unit test build Ubuntu18.04
* add missing dependency to build docu

+++ ad_map_access/impl/tests/route/RoutePlanningTests.cpp
@@ -215,7 +215,7 @@ void RoutePlanningTest::validateRouteConnections(route::FullRoute const &route, 
                   - route.roadSegments[i].drivableLaneSegments[0].routeLaneOffset)
         << "i:" << i << " j:" << j << " route:" << route;
 
-      if (j != static_cast<size_t>(route.roadSegments[i].drivableLaneSegments[j].routeLaneOffset))
+      if (static_cast<int32_t>(j) != route.roadSegments[i].drivableLaneSegments[j].routeLaneOffset)
       {
         EXPECT_NE(RouteIsConstant, routeBehavior);
       }
